% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{comparison_plot}
\alias{comparison_plot}
\title{Comparsion Plot}
\usage{
comparison_plot(X, y, model_list)
}
\description{
Plots the fitted functions obtained with various regressors (using their default values) on the
one-dimensional dataset (X, y).
}
