% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{data.algerianForestFire}
\alias{data.algerianForestFire}
\title{Algerian Forest Fire Dataset}
\usage{
data.algerianForestFire()
}
\description{
243 x 11. Reference: \url{https://archive.ics.uci.edu/ml/datasets/Algerian+Forest+Fires+Dataset++}
}
