% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{data.insurance}
\alias{data.insurance}
\title{Insurance Dataset}
\usage{
data.insurance()
}
\description{
5822 x 39. Reference: \url{https://archive.ics.uci.edu/ml/datasets/Insurance+Company+Benchmark+\%28COIL+2000\%29}
}
