% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{data.superconduct}
\alias{data.superconduct}
\title{Superconduct Dataset}
\usage{
data.superconduct()
}
\description{
21623 x 82. Reference: \url{https://archive.ics.uci.edu/ml/datasets/Superconductivty+Data}
}
