% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lessTest.R
\name{LESSBase}
\alias{LESSBase}
\title{Fit function: All data is used with the global estimator (no validation)
Tree method is used (no clustering)}
\description{
Fit function: All data is used with the global estimator (no validation)
Tree method is used (no clustering)

Fit function: All data is used with the global estimator (no validation)
Tree method is used (no clustering)
}
\section{Super class}{
\code{\link[lessTest:SklearnEstimator]{lessTest::SklearnEstimator}} -> \code{LESSBase}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LESSBase-new}{\code{LESSBase$new()}}
\item \href{#method-LESSBase-set_local_attributes}{\code{LESSBase$set_local_attributes()}}
\item \href{#method-LESSBase-fitnoval}{\code{LESSBase$fitnoval()}}
\item \href{#method-LESSBase-clone}{\code{LESSBase$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="lessTest" data-topic="SklearnEstimator" data-id="fit"><a href='../../lessTest/html/SklearnEstimator.html#method-SklearnEstimator-fit'><code>lessTest::SklearnEstimator$fit()</code></a></li>
<li><span class="pkg-link" data-pkg="lessTest" data-topic="SklearnEstimator" data-id="predict"><a href='../../lessTest/html/SklearnEstimator.html#method-SklearnEstimator-predict'><code>lessTest::SklearnEstimator$predict()</code></a></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LESSBase-new"></a>}}
\if{latex}{\out{\hypertarget{method-LESSBase-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LESSBase$new(replication = NA, scobject = NA, isFitted)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LESSBase-set_local_attributes"></a>}}
\if{latex}{\out{\hypertarget{method-LESSBase-set_local_attributes}{}}}
\subsection{Method \code{set_local_attributes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LESSBase$set_local_attributes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LESSBase-fitnoval"></a>}}
\if{latex}{\out{\hypertarget{method-LESSBase-fitnoval}{}}}
\subsection{Method \code{fitnoval()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LESSBase$fitnoval(X, y)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LESSBase-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LESSBase-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LESSBase$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
