% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/less.R
\name{LESSRegressor}
\alias{LESSRegressor}
\title{LESSRegressor}
\description{
LESSRegressor

LESSRegressor
}
\section{Super classes}{
\code{\link[less:BaseEstimator]{less::BaseEstimator}} -> \code{\link[less:SklearnEstimator]{less::SklearnEstimator}} -> \code{\link[less:LESSBase]{less::LESSBase}} -> \code{LESSRegressor}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LESSRegressor-new}{\code{LESSRegressor$new()}}
\item \href{#method-LESSRegressor-fit}{\code{LESSRegressor$fit()}}
\item \href{#method-LESSRegressor-predict}{\code{LESSRegressor$predict()}}
\item \href{#method-LESSRegressor-clone}{\code{LESSRegressor$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="less" data-topic="BaseEstimator" data-id="get_attributes"><a href='../../less/html/BaseEstimator.html#method-BaseEstimator-get_attributes'><code>less::BaseEstimator$get_attributes()</code></a></li>
<li><span class="pkg-link" data-pkg="less" data-topic="BaseEstimator" data-id="public_fields"><a href='../../less/html/BaseEstimator.html#method-BaseEstimator-public_fields'><code>less::BaseEstimator$public_fields()</code></a></li>
<li><span class="pkg-link" data-pkg="less" data-topic="BaseEstimator" data-id="set_random_state"><a href='../../less/html/BaseEstimator.html#method-BaseEstimator-set_random_state'><code>less::BaseEstimator$set_random_state()</code></a></li>
<li><span class="pkg-link" data-pkg="less" data-topic="LESSBase" data-id="check_input"><a href='../../less/html/LESSBase.html#method-LESSBase-check_input'><code>less::LESSBase$check_input()</code></a></li>
<li><span class="pkg-link" data-pkg="less" data-topic="LESSBase" data-id="fitnoval"><a href='../../less/html/LESSBase.html#method-LESSBase-fitnoval'><code>less::LESSBase$fitnoval()</code></a></li>
<li><span class="pkg-link" data-pkg="less" data-topic="LESSBase" data-id="fitnovalc"><a href='../../less/html/LESSBase.html#method-LESSBase-fitnovalc'><code>less::LESSBase$fitnovalc()</code></a></li>
<li><span class="pkg-link" data-pkg="less" data-topic="LESSBase" data-id="fitval"><a href='../../less/html/LESSBase.html#method-LESSBase-fitval'><code>less::LESSBase$fitval()</code></a></li>
<li><span class="pkg-link" data-pkg="less" data-topic="LESSBase" data-id="fitvalc"><a href='../../less/html/LESSBase.html#method-LESSBase-fitvalc'><code>less::LESSBase$fitvalc()</code></a></li>
<li><span class="pkg-link" data-pkg="less" data-topic="LESSBase" data-id="get_d_normalize"><a href='../../less/html/LESSBase.html#method-LESSBase-get_d_normalize'><code>less::LESSBase$get_d_normalize()</code></a></li>
<li><span class="pkg-link" data-pkg="less" data-topic="LESSBase" data-id="get_frac"><a href='../../less/html/LESSBase.html#method-LESSBase-get_frac'><code>less::LESSBase$get_frac()</code></a></li>
<li><span class="pkg-link" data-pkg="less" data-topic="LESSBase" data-id="get_n_neighbors"><a href='../../less/html/LESSBase.html#method-LESSBase-get_n_neighbors'><code>less::LESSBase$get_n_neighbors()</code></a></li>
<li><span class="pkg-link" data-pkg="less" data-topic="LESSBase" data-id="get_n_replications"><a href='../../less/html/LESSBase.html#method-LESSBase-get_n_replications'><code>less::LESSBase$get_n_replications()</code></a></li>
<li><span class="pkg-link" data-pkg="less" data-topic="LESSBase" data-id="get_n_subsets"><a href='../../less/html/LESSBase.html#method-LESSBase-get_n_subsets'><code>less::LESSBase$get_n_subsets()</code></a></li>
<li><span class="pkg-link" data-pkg="less" data-topic="LESSBase" data-id="get_random_state"><a href='../../less/html/LESSBase.html#method-LESSBase-get_random_state'><code>less::LESSBase$get_random_state()</code></a></li>
<li><span class="pkg-link" data-pkg="less" data-topic="LESSBase" data-id="get_scaling"><a href='../../less/html/LESSBase.html#method-LESSBase-get_scaling'><code>less::LESSBase$get_scaling()</code></a></li>
<li><span class="pkg-link" data-pkg="less" data-topic="LESSBase" data-id="get_val_size"><a href='../../less/html/LESSBase.html#method-LESSBase-get_val_size'><code>less::LESSBase$get_val_size()</code></a></li>
<li><span class="pkg-link" data-pkg="less" data-topic="LESSBase" data-id="print"><a href='../../less/html/LESSBase.html#method-LESSBase-print'><code>less::LESSBase$print()</code></a></li>
<li><span class="pkg-link" data-pkg="less" data-topic="LESSBase" data-id="set_local_attributes"><a href='../../less/html/LESSBase.html#method-LESSBase-set_local_attributes'><code>less::LESSBase$set_local_attributes()</code></a></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LESSRegressor-new"></a>}}
\if{latex}{\out{\hypertarget{method-LESSRegressor-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LESSRegressor$new(
  frac = NULL,
  n_replications = 20,
  random_state = NULL,
  n_subsets = NULL,
  n_neighbors = NULL,
  local_estimator = LinearRegression$new(),
  d_normalize = TRUE,
  global_estimator = DecisionTreeRegressor$new(),
  scaling = TRUE,
  cluster_method = NULL,
  distance_function = NULL,
  warnings = TRUE,
  val_size = NULL,
  tree_method = function(X) KDTree$new(X)
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LESSRegressor-fit"></a>}}
\if{latex}{\out{\hypertarget{method-LESSRegressor-fit}{}}}
\subsection{Method \code{fit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LESSRegressor$fit(X, y)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LESSRegressor-predict"></a>}}
\if{latex}{\out{\hypertarget{method-LESSRegressor-predict}{}}}
\subsection{Method \code{predict()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LESSRegressor$predict(X0)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LESSRegressor-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LESSRegressor-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LESSRegressor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
