% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lessTest.R
\name{linReg}
\alias{linReg}
\title{Apply linear regression}
\usage{
linReg(
  x = c(151, 174, 138, 186, 128, 136, 179, 163, 152, 131),
  y = c(63, 81, 56, 91, 47, 57, 76, 72, 62, 48)
)
}
\arguments{
\item{x}{vector of x values, default c(151, 174, 138, 186, 128, 136, 179, 163, 152, 131)}

\item{y}{vector of y values, default c(63, 81, 56, 91, 47, 57, 76, 72, 62, 48)}
}
\value{
Prints the relation between y and x
}
\description{
Apply linear regression
}
\examples{
linReg()
}
