% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/less.R
\name{rbf}
\alias{rbf}
\title{RBF kernel - L2 norm
This is is used as the default distance function in LESS}
\usage{
rbf(data, center, coeff = 0.01)
}
\description{
RBF kernel - L2 norm
This is is used as the default distance function in LESS
}
