% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superconduct_data.R
\docType{data}
\name{superconduct}
\alias{superconduct}
\title{Superconductivity Data Set}
\format{
A dataframe with 21263 rows and 82 variables
}
\source{
Kam Ham idieh, \email{khamidieh@gmail.com}, University of Pennsylvania, Statistics
}
\usage{
superconduct
}
\description{
Two dataset contains data on 21263 superconductors and their relevant features.
}
\examples{
data(superconduct)
}
\keyword{datasets}
